Struct sdl2_window::Sdl2Window

pub struct Sdl2Window {
    pub window: Window,
    pub context: GLContext,
    pub sdl_context: Sdl,
    pub video_subsystem: VideoSubsystem,
    // some fields omitted
}

A window implemented by SDL2 back-end.

Fields
window: Window

SDL window handle.

context: GLContext

Allow dead code because this keeps track of the OpenGL context. Will be released on drop.

sdl_context: Sdl

SDL context.

video_subsystem: VideoSubsystem

Video subsystem.

Implementations

impl Sdl2Window

pub fn new(settings: &WindowSettings) -> Result<Self, String>

Creates a new game window for SDL2. This will initialize SDL and the video subsystem. You can retrieve both via the public fields on the Sdl2Window struct.

pub fn with_subsystem(
video_subsystem: VideoSubsystem,
settings: &WindowSettings
) -> Result<Self, String>

Creates a window with the supplied SDL Video subsystem.

pub fn init_joysticks(&mut self) -> Result<u32, String>

Initialize the joystick subsystem. Required before joystick input events will be returned. Returns the number available or error.